+++ pytket/conanfile.py
@@ -35,9 +35,12 @@ class pytketRecipe(ConanFile):
         cmake.install()
 
     def _get_tket_version(self) -> str:
+        tket_ver = os.getenv("TKET_VERSION")
+        if tket_ver is not None:
+            return tket_ver
         # TKET_VERSION is in the parent directory in the repo, but will be in
         # the current directory once the package has been exported to conan cache
-        for try_dir in [".", ".."]:
+        for try_dir in [os.curdir, os.pardir]:
             path = os.path.join(try_dir, "TKET_VERSION")
             if os.path.exists(path):
                 return load(self, path).strip()
@@ -63,6 +66,6 @@ class pytketRecipe(ConanFile):
         copy(
             self,
             "TKET_VERSION",
-            os.path.join(self.recipe_folder, ".."),
+            os.path.join(self.recipe_folder, os.pardir),
             self.export_folder,
         )