+++ tket/test/src/test_GreedyPauli.cpp
@@ -708,6 +708,10 @@ SCENARIO("Test GreedyPauliSimp for individual gates") {
       get_op_ptr(OpType::PhasedX, {0.15, 0.2}),
       get_op_ptr(OpType::PhasedX, {0.5, -0.5}),
       get_op_ptr(OpType::PhasedX, {0.2, 1}),
+      get_op_ptr(OpType::TK1, {0.25, 0.5, 0.75}),
+      get_op_ptr(OpType::TK1, {0.012, 0.78, 0.1249}),
+      get_op_ptr(OpType::TK1, {-0.25, 0.5, 0.5}),
+      get_op_ptr(OpType::TK1, {0.12385, 0.281, 0.595}),
       get_op_ptr(OpType::T),
       get_op_ptr(OpType::Tdg),
   };
@@ -766,6 +770,14 @@ SCENARIO("Test GreedyPauliSimp pass construction") {
     CompilationUnit cu(c);
     REQUIRE(!gen_greedy_pauli_simp(0.3, 0.5, 500, 500, 0, false, 100, true)
                  ->apply(cu));
+  }
+  GIVEN("A circuit with implicit Clifford merging") {
+    Circuit c(1);
+    c.add_op<unsigned>(OpType::Rz, 1.005, {0});
+    c.add_op<unsigned>(OpType::PhasedX, {1.50, 1.505}, {0});
+    CompilationUnit cu(c);
+    CHECK(gen_greedy_pauli_simp(0.3, 0.5)->apply(cu));
+    REQUIRE(test_unitary_comparison(c, cu.get_circ_ref(), true));
   }
 }